Fixes for Copilot Code Review
