+++ src/v1/helper.ts
@@ -120,27 +120,4 @@ export function loopupEnum(deviceModel: DeviceModel, decodedMonitor: any, key: a
   return deviceModel.enumName(key, decodedMonitor[key]);
 }
 
-export function normalizeBoolean(value: any): boolean {
-  if (typeof value === 'boolean') {
-    return value;
-  }
-  if (typeof value === 'number') {
-    return value === 1;
-  }
-  if (typeof value === 'string') {
-    const v = value.toLowerCase();
-    return v === '1' || v === 'true' || v === 'on';
-  }
-  return !!value;
-}
-
-export function normalizeNumber(value: any): number | null {
-  if (value === null || value === undefined) {
-    return null;
-  }
-  if (typeof value === 'number') {
-    return value;
-  }
-  const n = Number(value);
-  return Number.isNaN(n) ? null : n;
-}
+export { normalizeBoolean, normalizeNumber } from '../utils/normalize.js';
